v1_monolithic/requirements.txt
# Core Data Engineering
apache-airflow==2.7.3
dbt-core==1.6.0
dbt-snowflake==1.6.0
pandas==2.1.0
numpy==1.24.3
pyspark==3.4.1

# Streaming & Messaging
kafka-python==2.0.2
confluent-kafka==2.3.0
pyspark-streaming==3.4.1

# Database & Storage
snowflake-connector-python==3.0.0
psycopg2-binary==2.9.7
pymongo==4.5.0
boto3==1.28.0
s3fs==2023.6.0

# Data Quality & Monitoring
great-expectations==0.17.0
pandera==0.17.0
datadog==0.48.0
prometheus-client==0.17.1

# Utilities
pydantic==2.3.0
python-dotenv==1.0.0
pyyaml==6.0.1
click==8.1.7
requests==2.31.0
schedule==1.2.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1

# Development
black==23.7.0
flake8==6.0.0
mypy==1.5.1
pre-commit==3.3.3
